Important Notice: This project has transitioned from the MIT License to the GNU General Public License v3.0 (GPLv3).
All copies of the software downloaded or obtained before this change remain licensed under the MIT License. This means that your rights under that license are fully preserved.

This license update applies only to this version and all future versions of the software from this point forward.

We respect and uphold your legal rights to use, modify, and distribute any previous versions released under the MIT License.